Install script rewrite Modified all instances of yes/no choices to reflect that "no" is the default option.
rewrote some of the verbage in config.sh to clarify the scope of the 
input

simplified the choosing of the facade-worker directory. The existing 
conditional was ineffectual, because either option was capable of 
choosing a pre-existing directory. It was made clear in the verbage that 
if you choose an alredy existing directory, it will be used, otherwise a 
new directory will be created.